Use "sample" instead of "question"
A lot of the prompts to the model aren't actually questions. And for
those that are, it was confusing to have a question field inside the
Question class.

We still use the term "question" for one of the prompt formats. It's
also technically incorrect, but it's what Ganguli et al. use in their
paper.