Refactor eval.py into multiple modules
I felt it was getting rather long as a single file.

Logically, I'm not sure this is the best way to split things up. I
originally wanted to split the code into requests.py and results.py.
requests.py would contain everything up to and including the API calls.
And results.py would contain the code to save the output to a file.
However, there wasn't a clean way to divide the classes and functions: I
would have ended up with circular imports, which are a sign that the
division is a poor one.

Splitting classes from functions divides things according to what they
are, not so much how they're used. But at least it avoid circular
imports and is relatively unambiguous.