Format prompts so they look better alongside code
But strip out the indentations (and optionally newlines). Note that this
is done when making the unformatted string, not the final prompt sent to
the model. This way is more efficient (since it happens a handful of
times, not for each of the thousands of samples). It also preserves all
whitespace inserted by the formatting step, for better or worse.