Truncate model answer to make grading more accurate
There were a lot of cases where model's answer was along the lines of
    "yes, there is no reason to deny admission"
Until now, that was graded as UNKNOWN since it contained both "yes"
and "no." It should be graded as CORRECT.

Truncating the answer feels like a bit of a hack in some respects. But
it's straightforward and very effective (at least for these examples).