Change all relative imports to absolute
This requires one to be in the root directory of the repository to run
anything. But even before this change, I believe there were
restrictions.

Something this unlocks is the ability to run files as scripts:
    python path/to/module.py
Until now, several files could only be run as modules:
    python -m path.to.module
Now both forms work (again assuming you're in the root directory of the
repo).